pyproject.toml
[project]
authors = [
    {name = "Andrey Solovyev", email = "solovyov.net@yandex.ru"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Вычислитель отличий"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-cov>=6.1.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.9.5",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
